Move getThemeShowcaseTitle
